requirements.txt
Django>=1.8,<2.0
psycopg2
django-querycount
uwsgi
django-haystack==2.6
elasticsearch==1.8
